api/requirements.txt
django==2.0.6
djangorestframework==3.8.2
drf-haystack==1.8.3
